fixed update

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -1,14 +1,14 @@
 #include "lists.h"
 /**
- * print_listint - This function will print all the elements of linked list.
+ * print_listint - This function will
+ * print all the elements of linked list.
  * @h: This is our data type to the  pointer of the struct.
- * @rep: my counter
  *
  * Return: The elements of the str rep
  */
 size_t print_listint(const listint_t *h)
 {
-	unsigned int rep = 0;
+	size_t rep = 0;
 	const listint_t	 *mover = h;
 
 	while (mover != NULL)
